+++ packages/aesthetic-adapter-typestyle/src/NativeAdapter.js
@@ -5,12 +5,18 @@
  */
 
 import { Adapter } from 'aesthetic';
-import { style } from 'typestyle';
+import { TypeStyle } from 'typestyle';
 
-import type { StyleSheet } from '../../types';
+import type { Statement, StyleSheet } from '../../types';
 
 export default class TypeStyleAdapter extends Adapter {
-  transform<T: Object>(styleName: string, statement: T): StyleSheet {
+  constructor(options?: Object = {}) {
+    super(options);
+
+    this.typeStyle = new TypeStyle({ autoGenerateTag: true });
+  }
+
+  transform(styleName: string, statement: Statement): StyleSheet {
     if (__DEV__) {
       if (this.native) {
         throw new Error('TypeStyle does not support React Native.');
@@ -22,7 +28,7 @@ export default class TypeStyleAdapter extends Adapter {
     Object.keys(statement).forEach((selector) => {
       const value = statement[selector];
 
-      output[selector] = (typeof value === 'string') ? value : style(value);
+      output[selector] = (typeof value === 'string') ? value : this.typeStyle.style(value);
     });
 
     return output;